+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Removed currying from `assign` based on new strategy, will need to re approach this perhaps?
 - Re wrote `isEqual` again, making it far more flexible again as well as far more performant
 - `range` will not return a type error if given a string
+- Renamed `isEqual` to `deepEq` to match other `eq` functions
 
 ### Improved